+++ src/bling/LT.java
@@ -7,7 +7,7 @@ public class LT extends RelOp {
 	}
 
 	@Override
-	protected Boolean compute(Integer l, Integer r) {
+	protected Boolean compute(Long l, Long r) {
 		return l < r;
 	}
 }
